+++ src/android/FirebasePlugin.java
@@ -34,6 +34,7 @@ import com.google.android.gms.auth.api.signin.GoogleSignInOptions;
 import com.google.android.gms.common.api.ApiException;
 import com.google.android.gms.common.api.CommonStatusCodes;
 import com.google.android.gms.common.api.GoogleApiClient;
+import com.google.android.gms.tasks.Continuation;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
@@ -61,7 +62,12 @@ import com.google.firebase.firestore.Query;
 import com.google.firebase.firestore.QueryDocumentSnapshot;
 import com.google.firebase.firestore.QuerySnapshot;
 import com.google.firebase.firestore.Query.Direction;
+import com.google.firebase.functions.FirebaseFunctions;
+import com.google.firebase.functions.FirebaseFunctionsException;
+import com.google.firebase.functions.HttpsCallableResult;
 import com.google.firebase.iid.FirebaseInstanceId;
+import com.google.firebase.installations.FirebaseInstallations;
+import com.google.firebase.installations.InstallationTokenResult;
 import com.google.firebase.messaging.FirebaseMessaging;
 import com.google.firebase.remoteconfig.FirebaseRemoteConfig;
 import com.google.firebase.remoteconfig.FirebaseRemoteConfigInfo;
@@ -98,6 +104,11 @@ import com.google.firebase.FirebaseTooManyRequestsException;
 import com.google.firebase.auth.PhoneAuthCredential;
 import com.google.firebase.auth.PhoneAuthProvider;
 import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+import com.google.gson.JsonElement;
+import com.google.gson.JsonPrimitive;
+import com.google.gson.JsonSerializationContext;
+import com.google.gson.JsonSerializer;
 import com.google.gson.reflect.TypeToken;
 
 import static android.content.Context.MODE_PRIVATE;
@@ -108,6 +119,7 @@ public class FirebasePlugin extends CordovaPlugin {
     private FirebaseAnalytics mFirebaseAnalytics;
     private FirebaseCrashlytics firebaseCrashlytics;
     private FirebaseFirestore firestore;
+    private FirebaseFunctions functions;
     private Gson gson;
     private FirebaseAuth.AuthStateListener authStateListener;
     private boolean authStateChangeListenerInitialized = false;
@@ -171,8 +183,18 @@ public class FirebasePlugin extends CordovaPlugin {
                     authStateListener = new AuthStateListener();
                     FirebaseAuth.getInstance().addAuthStateListener(authStateListener);
 
-                    firestore = FirebaseFirestore.getInstance();
-                    gson = new Gson();
+                    firestore = FirebaseFirestore.getInstance();                  
+                    functions = FirebaseFunctions.getInstance();
+
+                    gson = new GsonBuilder()
+                    .registerTypeAdapter(Double.class, new JsonSerializer<Double>() {
+                        public JsonElement serialize(Double src, Type typeOfSrc, JsonSerializationContext context) {
+                            if (src.isNaN() || src.isInfinite())
+                                return new JsonPrimitive(src.toString());
+                            return new JsonPrimitive(src);
+                        }
+                    })
+                    .create();
 
                     if (extras != null && extras.size() > 1) {
                         if (FirebasePlugin.notificationStack == null) {
@@ -355,13 +377,21 @@ public class FirebasePlugin extends CordovaPlugin {
                 this.listenToFirestoreCollection(args, callbackContext);
             } else if (action.equals("removeFirestoreListener")) {
                 this.removeFirestoreListener(args, callbackContext);
+            } else if (action.equals("functionsHttpsCallable")) {
+                this.functionsHttpsCallable(args, callbackContext);
             } else if (action.equals("grantPermission")
                     || action.equals("setBadgeNumber")
                     || action.equals("getBadgeNumber")
                     ) {
                 // Stubs for other platform methods
                 callbackContext.sendPluginResult(new PluginResult(PluginResult.Status.OK, true));
-            }else{
+            } else if (action.equals("deleteInstallationId")) {
+                this.deleteInstallationId(args, callbackContext);
+            } else if (action.equals("getInstallationId")) {
+                this.getInstallationId(args, callbackContext);
+            } else if (action.equals("getInstallationToken")) {
+                this.getInstallationToken(args, callbackContext);
+            } else{
                 callbackContext.error("Invalid action: " + action);
                 return false;
             }
@@ -649,20 +679,42 @@ public class FirebasePlugin extends CordovaPlugin {
         });
     }
 
-    private void logEvent(final CallbackContext callbackContext, final String name, final JSONObject params)
-            throws JSONException {
+    private Bundle createBundleFromJSONObject(final JSONObject params) throws JSONException {
         final Bundle bundle = new Bundle();
-        Iterator iter = params.keys();
+        Iterator<String> iter = params.keys();
         while (iter.hasNext()) {
-            String key = (String) iter.next();
-            Object value = params.get(key);
-
-            if (value instanceof Integer || value instanceof Double) {
-                bundle.putFloat(key, ((Number) value).floatValue());
+            String key = iter.next();
+            Object obj = params.get(key);
+            if (obj instanceof Integer) {
+                bundle.putInt(key, (Integer) obj);
+            } else if (obj instanceof Double) {
+                bundle.putDouble(key, (Double) obj);
+            } else if (obj instanceof Float) {
+                bundle.putFloat(key, (Float) obj);
+            } else if (obj instanceof JSONObject) {
+                Bundle item = this.createBundleFromJSONObject((JSONObject) obj);
+                bundle.putBundle(key, item);
+            } else if (obj instanceof JSONArray) {
+                JSONArray objArr = (JSONArray) obj;
+                ArrayList<Bundle> bundleArray = new ArrayList<Bundle>(objArr.length());
+                for (int idx = 0; idx < objArr.length(); idx++) {
+                    Object tmp = objArr.get(idx);
+                    if (tmp instanceof JSONObject) {
+                        Bundle item = createBundleFromJSONObject(objArr.getJSONObject(idx));
+                        bundleArray.add(item);
+                    }
+                }
+                bundle.putParcelableArrayList(key, bundleArray);
             } else {
-                bundle.putString(key, value.toString());
+                bundle.putString(key, obj.toString());
             }
         }
+        return bundle;
+    }
+
+    private void logEvent(final CallbackContext callbackContext, final String name, final JSONObject params)
+            throws JSONException {
+        final Bundle bundle = this.createBundleFromJSONObject(params);
 
         cordova.getThreadPool().execute(new Runnable() {
             public void run() {
@@ -857,7 +909,7 @@ public class FirebasePlugin extends CordovaPlugin {
         cordova.getThreadPool().execute(new Runnable() {
             public void run() {
                 try {
-                    handleBooleanTaskOutcome(FirebaseRemoteConfig.getInstance().activate(), callbackContext);
+                    handleTaskOutcomeWithBooleanResult(FirebaseRemoteConfig.getInstance().activate(), callbackContext);
                 } catch (Exception e) {
                     handleExceptionWithContext(e, callbackContext);
                 }
@@ -869,7 +921,7 @@ public class FirebasePlugin extends CordovaPlugin {
         cordova.getThreadPool().execute(new Runnable() {
             public void run() {
                 try {
-                    handleBooleanTaskOutcome(FirebaseRemoteConfig.getInstance().fetchAndActivate(), callbackContext);
+                    handleTaskOutcomeWithBooleanResult(FirebaseRemoteConfig.getInstance().fetchAndActivate(), callbackContext);
                 } catch (Exception e) {
                     handleExceptionWithContext(e, callbackContext);
                 }
@@ -910,7 +962,6 @@ public class FirebasePlugin extends CordovaPlugin {
                     JSONObject info = new JSONObject();
 
                     JSONObject settings = new JSONObject();
-                    settings.put("developerModeEnabled", remoteConfigInfo.getConfigSettings().isDeveloperModeEnabled());
                     info.put("configSettings", settings);
 
                     info.put("fetchTimeMillis", remoteConfigInfo.getFetchTimeMillis());
@@ -943,7 +994,7 @@ public class FirebasePlugin extends CordovaPlugin {
             }
         });
     }
-	
+
 	private void setConfigSettings(final CallbackContext callbackContext, final JSONArray args) throws JSONException {
         cordova.getThreadPool().execute(new Runnable() {
             public void run() {
@@ -1131,6 +1182,14 @@ public class FirebasePlugin extends CordovaPlugin {
                 } catch (Exception e) {
                     handleExceptionWithContext(e, callbackContext);
                 }
+            }
+
+        }).addOnFailureListener(new OnFailureListener() {
+            @Override
+            public void onFailure(@NonNull Exception e) {
+                // Something went wrong getting ID token so return other user data
+                callbackContext.sendPluginResult(new PluginResult(PluginResult.Status.OK, returnResults));
+                handleExceptionWithoutContext(e);
             }
         });
     }
@@ -1478,7 +1537,7 @@ public class FirebasePlugin extends CordovaPlugin {
             }
         });
     }
-    
+
     public void createUserWithEmailAndPassword(final CallbackContext callbackContext, final JSONArray args){
         cordova.getThreadPool().execute(new Runnable() {
             public void run() {
@@ -2518,6 +2577,120 @@ public class FirebasePlugin extends CordovaPlugin {
         return removed;
     }
 
+    //
+    // Functions
+    //
+    private void functionsHttpsCallable(JSONArray args, CallbackContext callbackContext) throws JSONException {
+        cordova.getThreadPool().execute(new Runnable() {
+            public void run() {
+                try {
+                    String name = args.getString(0);
+                    functions.getHttpsCallable(name)
+                        .call(args.get(1))
+                        .addOnSuccessListener(new OnSuccessListener<HttpsCallableResult>() {
+                            @Override
+                            public void onSuccess(HttpsCallableResult httpsCallableResult) {
+                                try{
+                                    if (httpsCallableResult.getData() instanceof Map) {
+                                        callbackContext.success(mapToJsonObject((Map<String, Object>) httpsCallableResult.getData()));
+                                    } else if (httpsCallableResult.getData() instanceof ArrayList) {
+                                        callbackContext.success(objectToJsonArray(httpsCallableResult.getData()));
+                                    } else if (httpsCallableResult.getData() instanceof Integer) {
+                                        callbackContext.success((int) httpsCallableResult.getData());
+                                    } else if (httpsCallableResult.getData() instanceof String) {
+                                        callbackContext.success((String) httpsCallableResult.getData());
+                                    } else {
+                                        callbackContext.success((byte[]) httpsCallableResult.getData());
+                                    }
+                                } catch (Exception e){
+                                    handleExceptionWithContext(e, callbackContext);
+                                }
+                            }
+                        })
+                        .addOnFailureListener(new OnFailureListener() {
+                            @Override
+                            public void onFailure(@NonNull Exception e) {
+                                if (e instanceof FirebaseFunctionsException) {
+                                    this.onFailure((FirebaseFunctionsException) e);
+                                    return;
+                                }
+                                handleExceptionWithContext(e, callbackContext);
+                            }
+
+                            void onFailure(@NonNull FirebaseFunctionsException e) {
+                                if (e.getDetails() == null) {
+                                    handleExceptionWithContext(e, callbackContext);
+                                    return;
+                                }
+                                if (e.getDetails() instanceof String) {
+                                    callbackContext.error(e.getDetails().toString());
+                                } else {
+                                    try {
+                                        callbackContext.error(mapToJsonObject((Map<String, Object>) e.getDetails()));
+                                    } catch (JSONException ex) {
+                                        handleExceptionWithContext(ex, callbackContext);
+                                    }
+                                }
+                            }
+                        });
+                } catch (Exception e) {
+                    handleExceptionWithContext(e, callbackContext);
+                }
+            }
+        });
+    }
+
+    //
+    // Installations
+    //
+    private void deleteInstallationId(JSONArray args, CallbackContext callbackContext) throws JSONException {
+        cordova.getThreadPool().execute(new Runnable() {
+            public void run() {
+                try {
+                    handleTaskOutcome(FirebaseInstallations.getInstance().delete(), callbackContext);
+                } catch (Exception e) {
+                    handleExceptionWithContext(e, callbackContext);
+                }
+            }
+        });
+    }
+
+    private void getInstallationId(JSONArray args, CallbackContext callbackContext) throws JSONException {
+        cordova.getThreadPool().execute(new Runnable() {
+            public void run() {
+                try {
+                    handleTaskOutcomeWithStringResult(FirebaseInstallations.getInstance().getId(), callbackContext);
+                } catch (Exception e) {
+                    handleExceptionWithContext(e, callbackContext);
+                }
+            }
+        });
+    }
+
+    private void getInstallationToken(JSONArray args, CallbackContext callbackContext) throws JSONException {
+        cordova.getThreadPool().execute(new Runnable() {
+            public void run() {
+                try {
+                    FirebaseInstallations.getInstance().getToken(/* forceRefresh */true)
+                            .addOnCompleteListener(new OnCompleteListener<InstallationTokenResult>() {
+                                @Override
+                                public void onComplete(@NonNull Task<InstallationTokenResult> task) {
+                                    if (task.isSuccessful() || task.getException() == null) {
+                                        callbackContext.success(task.getResult().getToken());
+                                    }else if(task.getException() != null){
+                                        callbackContext.error(task.getException().getMessage());
+                                    }else{
+                                        callbackContext.error("Task failed for unknown reason");
+                                    }
+                                }
+                            });
+                } catch (Exception e) {
+                    handleExceptionWithContext(e, callbackContext);
+                }
+            }
+        });
+    }
+
 
     /*
      * Helper methods
@@ -2647,7 +2820,7 @@ public class FirebasePlugin extends CordovaPlugin {
         }
     }
 
-    private void handleBooleanTaskOutcome(@NonNull Task<Boolean> task, CallbackContext callbackContext) {
+    private void handleTaskOutcomeWithBooleanResult(@NonNull Task<Boolean> task, CallbackContext callbackContext) {
         try {
             task.addOnCompleteListener(new OnCompleteListener<Boolean>() {
                 @Override
@@ -2655,6 +2828,29 @@ public class FirebasePlugin extends CordovaPlugin {
                     try {
                         if (task.isSuccessful() || task.getException() == null) {
                             callbackContext.success(conformBooleanForPluginResult(task.getResult()));
+                        }else if(task.getException() != null){
+                            callbackContext.error(task.getException().getMessage());
+                        }else{
+                            callbackContext.error("Task failed for unknown reason");
+                        }
+                    } catch (Exception e) {
+                        handleExceptionWithContext(e, callbackContext);
+                    }
+                };
+            });
+        } catch (Exception e) {
+            handleExceptionWithContext(e, callbackContext);
+        }
+    }
+
+    private void handleTaskOutcomeWithStringResult(@NonNull Task<String> task, CallbackContext callbackContext) {
+        try {
+            task.addOnCompleteListener(new OnCompleteListener<String>() {
+                @Override
+                public void onComplete(@NonNull Task<String> task) {
+                    try {
+                        if (task.isSuccessful() || task.getException() == null) {
+                            callbackContext.success(task.getResult());
                         }else if(task.getException() != null){
                             callbackContext.error(task.getException().getMessage());
                         }else{
@@ -2766,6 +2962,16 @@ public class FirebasePlugin extends CordovaPlugin {
     private JSONObject mapToJsonObject(Map<String, Object> map) throws JSONException {
         String jsonString = gson.toJson(map);
         return new JSONObject(jsonString);
+    }
+
+    private JSONObject objectToJsonObject(Object object) throws JSONException {
+        String jsonString = gson.toJson(object);
+        return new JSONObject(jsonString);
+    }
+
+    private JSONArray objectToJsonArray(Object object) throws JSONException {
+        String jsonString = gson.toJson(object);
+        return new JSONArray(jsonString);
     }
 
     private void logMessageToCrashlytics(String message){