+++ src/ios/FirebasePlugin.m
@@ -10,6 +10,7 @@
 @import FirebaseRemoteConfig;
 @import FirebasePerformance;
 @import FirebaseAuth;
+@import FirebaseFunctions;
 @import UserNotifications;
 @import CommonCrypto;
 @import AuthenticationServices;
@@ -41,6 +42,7 @@ static FIRFirestore* firestore;
 static NSUserDefaults* preferences;
 static NSDictionary* googlePlist;
 static NSMutableDictionary* firestoreListeners;
+static NSString* currentInstallationId;
 
 
 + (FirebasePlugin*) firebasePlugin {
@@ -75,7 +77,7 @@ static NSMutableDictionary* firestoreListeners;
         if([self getGooglePlistFlagWithDefaultValue:FIREBASE_PERFORMANCE_COLLECTION_ENABLED defaultValue:YES]){
             [self setPreferenceFlag:FIREBASE_PERFORMANCE_COLLECTION_ENABLED flag:YES];
         }
-        
+
         // Set actionable categories if pn-actions.json exist in bundle
         [self setActionableNotifications];
 
@@ -86,6 +88,7 @@ static NSMutableDictionary* firestoreListeners;
 
         authCredentials = [[NSMutableDictionary alloc] init];
         firestoreListeners = [[NSMutableDictionary alloc] init];
+        self.traces = [NSMutableDictionary new];
     }@catch (NSException *exception) {
         [self handlePluginExceptionWithoutContext:exception];
     }
@@ -113,7 +116,7 @@ static NSMutableDictionary* firestoreListeners;
                 NSString *actionId = [action objectForKey:@"id"];
                 NSString *actionTitle = [action objectForKey:@"title"];
                 UNNotificationActionOptions options = UNNotificationActionOptionNone;
-                
+
                 id mode = [action objectForKey:@"foreground"];
                 if (mode != nil && (([mode isKindOfClass:[NSString class]] && [mode isEqualToString:@"true"]) || [mode boolValue])) {
                     options |= UNNotificationActionOptionForeground;
@@ -122,7 +125,7 @@ static NSMutableDictionary* firestoreListeners;
                 if (destructive != nil && (([destructive isKindOfClass:[NSString class]] && [destructive isEqualToString:@"true"]) || [destructive boolValue])) {
                     options |= UNNotificationActionOptionDestructive;
                 }
-                
+
                 [buttons addObject:[UNNotificationAction actionWithIdentifier:actionId
                     title:NSLocalizedString(actionTitle, nil) options:options]];
             }
@@ -133,6 +136,16 @@ static NSMutableDictionary* firestoreListeners;
 
         // Initialize categories
         [[UNUserNotificationCenter currentNotificationCenter] setNotificationCategories:categories];
+        
+        // Initialize installation ID change listner
+        __weak __auto_type weakSelf = self;
+        self.installationIDObserver = [[NSNotificationCenter defaultCenter]
+                addObserverForName: FIRInstallationIDDidChangeNotification
+                            object:nil
+                             queue:nil
+                        usingBlock:^(NSNotification * _Nonnull notification) {
+            [weakSelf sendNewInstallationId];
+        }];
     }@catch (NSException *exception) {
         [self handlePluginExceptionWithoutContext:exception];
     }
@@ -144,6 +157,9 @@ static NSMutableDictionary* firestoreListeners;
     [[GIDSignIn sharedInstance] handleURL:url];
 }
 
+/*************************************************/
+#pragma mark - plugin API
+/*************************************************/
 - (void)setAutoInitEnabled:(CDVInvokedUrlCommand *)command {
     @try {
         bool enabled = [[command.arguments objectAtIndex:0] boolValue];
@@ -310,7 +326,7 @@ static NSMutableDictionary* firestoreListeners;
                                 [self registerForRemoteNotifications];
                             }
                             [self handleBoolResultWithPotentialError:error command:command result:granted];
-                            
+
                         }@catch (NSException *exception) {
                             [self handlePluginExceptionWithContext:exception :command];
                         }
@@ -630,7 +646,7 @@ static NSMutableDictionary* firestoreListeners;
         NSString* password = [command.arguments objectAtIndex:1];
         FIRAuthCredential* authCredential = [FIREmailAuthProvider credentialWithEmail:email password:password];
         NSNumber* key = [self saveAuthCredential:authCredential];
-        
+
         NSMutableDictionary* result = [[NSMutableDictionary alloc] init];
         [result setValue:@"true" forKey:@"instantVerification"];
         [result setValue:key forKey:@"id"];
@@ -835,9 +851,13 @@ static NSMutableDictionary* firestoreListeners;
     [userInfo setValue:user.uid forKey:@"uid"];
     [userInfo setValue:@(user.isAnonymous ? true : false) forKey:@"isAnonymous"];
     [user getIDTokenWithCompletion:^(NSString * _Nullable token, NSError * _Nullable error) {
-        [userInfo setValue:token forKey:@"idToken"];
+        if(error == nil){
+            [userInfo setValue:token forKey:@"idToken"];
+        }
         [user getIDTokenResultWithCompletion:^(FIRAuthTokenResult * _Nullable tokenResult, NSError * _Nullable error) {
-            [userInfo setValue:tokenResult.signInProvider forKey:@"providerId"];
+            if(error == nil){
+                [userInfo setValue:tokenResult.signInProvider forKey:@"providerId"];
+            }
             [self.commandDelegate sendPluginResult:[CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsDictionary:userInfo] callbackId:command.callbackId];
         }];
     }];
@@ -1174,12 +1194,13 @@ static NSMutableDictionary* firestoreListeners;
 
 - (void)logError:(CDVInvokedUrlCommand *)command {
     [self.commandDelegate runInBackground:^{
-        NSString* errorMessage = [command.arguments objectAtIndex:0];
-
-        CDVPluginResult *pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK];
         @try {
+            NSString* errorMessage = [command.arguments objectAtIndex:0];
+            CDVPluginResult *pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK];
             if(![self isCrashlyticsEnabled]){
                 pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_ERROR messageAsString:@"Cannot log error - Crashlytics collection is disabled"];
+            }else if([command.arguments objectAtIndex:0] == [NSNull null]){
+                pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_ERROR messageAsString:@"Cannot log error - error message is empty"];
             }
             // We can optionally be passed a stack trace from stackTrace.js which we'll put in userInfo.
             else if ([command.arguments count] > 1) {
@@ -1272,19 +1293,19 @@ static NSMutableDictionary* firestoreListeners;
     [self.commandDelegate runInBackground:^{
         @try {
             FIRRemoteConfig* remoteConfig = [FIRRemoteConfig remoteConfig];
-            
+
             FIRRemoteConfigSettings* settings = [[FIRRemoteConfigSettings alloc] init];
-            
+
             if([command.arguments objectAtIndex:0] != [NSNull null]){
                 settings.fetchTimeout = [[command.arguments objectAtIndex:0] longValue];
             }
-            
+
             if([command.arguments objectAtIndex:1] != [NSNull null]){
                 settings.minimumFetchInterval = [[command.arguments objectAtIndex:1] longValue];
             }
-            
+
             remoteConfig.configSettings = settings;
-            
+
             [self sendPluginSuccess:command];
         }@catch (NSException *exception) {
             [self handlePluginExceptionWithContext:exception :command];
@@ -1374,13 +1395,22 @@ static NSMutableDictionary* firestoreListeners;
     [self.commandDelegate runInBackground:^{
         @try {
             FIRRemoteConfig* remoteConfig = [FIRRemoteConfig remoteConfig];
-            NSArray* keys = [remoteConfig allKeysFromSource:FIRRemoteConfigSourceDefault];
+            NSArray* defaultKeys = [remoteConfig allKeysFromSource:FIRRemoteConfigSourceDefault];
+            NSArray* remoteKeys = [remoteConfig allKeysFromSource:FIRRemoteConfigSourceRemote];
+            NSArray* staticKeys = [remoteConfig allKeysFromSource:FIRRemoteConfigSourceStatic];
+            NSArray* keys = defaultKeys;
+            if([keys count] == 0){
+                keys = remoteKeys;
+            }
+            if([keys count] == 0){
+                keys = staticKeys;
+            }
             NSMutableDictionary* result = [[NSMutableDictionary alloc] init];
-            
+
             for (NSString* key in keys) {
                 [result setObject:remoteConfig[key].stringValue forKey:key];
             }
-    
+
             CDVPluginResult *pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsDictionary:result];
             [self.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
         }@catch (NSException *exception) {
@@ -1412,11 +1442,8 @@ static NSMutableDictionary* firestoreListeners;
             NSInteger fetchTimeout = remoteConfig.configSettings.fetchTimeout;
             NSDate* lastFetchTime = remoteConfig.lastFetchTime;
             FIRRemoteConfigFetchStatus lastFetchStatus = remoteConfig.lastFetchStatus;
-            // isDeveloperModeEnabled is deprecated new recommnded way to check is minimumFetchInterval == 0
-            BOOL isDeveloperModeEnabled = minimumFetchInterval == 0 ? true : false;
 
             NSDictionary* configSettings = @{
-                @"developerModeEnabled": [NSNumber numberWithBool:isDeveloperModeEnabled],
                 @"minimumFetchInterval": [NSNumber numberWithInteger:minimumFetchInterval],
                 @"fetchTimeout": [NSNumber numberWithInteger:fetchTimeout],
             };
@@ -1470,16 +1497,11 @@ static NSMutableDictionary* firestoreListeners;
     [self.commandDelegate runInBackground:^{
         @try {
             NSString* traceName = [command.arguments objectAtIndex:0];
-            FIRTrace *trace = [self.traces objectForKey:traceName];
-
-            if ( self.traces == nil) {
-                self.traces = [NSMutableDictionary new];
-            }
+            FIRTrace* trace = [self.traces objectForKey:traceName];
 
             if (trace == nil) {
                 trace = [FIRPerformance startTraceWithName:traceName];
                 [self.traces setObject:trace forKey:traceName ];
-
             }
 
             CDVPluginResult *pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK];
@@ -1682,7 +1704,7 @@ static NSMutableDictionary* firestoreListeners;
                     [self handlePluginExceptionWithContext:exception :command];
                 }
             }];
-            
+
             NSMutableDictionary* jsResult = [[NSMutableDictionary alloc] init];;
             [jsResult setObject:@"id" forKey:@"eventType"];
             NSNumber* key = [self saveFirestoreListener:listener];
@@ -1702,12 +1724,12 @@ static NSMutableDictionary* firestoreListeners;
             if([command.arguments objectAtIndex:1] != [NSNull null]){
                 filters = [command.arguments objectAtIndex:1];
             }
-            
+
             FIRQuery* query = [firestore collectionWithPath:collection];
             if(filters != nil){
                 query = [self applyFiltersToFirestoreCollectionQuery:filters query:query];
             }
-            
+
             [query getDocumentsWithCompletion:^(FIRQuerySnapshot * _Nullable snapshot, NSError * _Nullable error) {
                 if (error != nil) {
                     [self sendPluginErrorWithMessage:error.localizedDescription:command];
@@ -1734,12 +1756,12 @@ static NSMutableDictionary* firestoreListeners;
                 filters = [command.arguments objectAtIndex:1];
             }
             bool includeMetadata = [command.arguments objectAtIndex:2];
-            
+
             FIRQuery* query = [firestore collectionWithPath:collection];
             if(filters != nil){
                 query = [self applyFiltersToFirestoreCollectionQuery:filters query:query];
             }
-            
+
             id<FIRListenerRegistration> listener = [query
                 addSnapshotListenerWithIncludeMetadataChanges:includeMetadata
                 listener:^(FIRQuerySnapshot *snapshot, NSError *error) {
@@ -1747,7 +1769,7 @@ static NSMutableDictionary* firestoreListeners;
                         if(snapshot != nil){
                             NSMutableDictionary* jsResult = [[NSMutableDictionary alloc] init];
                             [jsResult setObject:@"change" forKey:@"eventType"];
-                            
+
                             NSMutableDictionary* documents = [[NSMutableDictionary alloc] init];
                             bool hasDocuments = false;
                             for (FIRDocumentChange* dc in snapshot.documentChanges) {
@@ -1841,12 +1863,12 @@ static NSMutableDictionary* firestoreListeners;
 
 - (id) getFilterValueAsType: (NSArray*)filter  valueIndex:(int)valueIndex typeIndex:(int)typeIndex{
     id typedValue = [filter objectAtIndex:valueIndex];
-    
+
     NSString* type = @"string";
     if([filter objectAtIndex:typeIndex] != nil){
         type = [filter objectAtIndex:typeIndex];
     }
-    
+
     if([type isEqual:@"boolean"]){
         if([typedValue isKindOfClass:[NSNumber class]]){
             typedValue = [NSNumber numberWithBool:typedValue];
@@ -1874,7 +1896,7 @@ static NSMutableDictionary* firestoreListeners;
             }
         }
     }
-    
+
     return typedValue;
 }
 
@@ -1887,10 +1909,12 @@ static NSMutableDictionary* firestoreListeners;
 }
 
 - (NSNumber*) saveFirestoreListener: (id<FIRListenerRegistration>) firestoreListener {
-    int id = [self generateId];
-    NSNumber* key = [NSNumber numberWithInt:id];
-    [firestoreListeners setObject:firestoreListener forKey:key];
-    return key;
+    @synchronized (firestoreListeners) {
+        int id = [self generateId];
+        NSNumber* key = [NSNumber numberWithInt:id];
+        [firestoreListeners setObject:firestoreListener forKey:key];
+        return key;
+    }
 }
 
 - (void) removeFirestoreListener:(CDVInvokedUrlCommand*)command {
@@ -1910,19 +1934,105 @@ static NSMutableDictionary* firestoreListeners;
 }
 
 - (bool) _removeFirestoreListener: (NSNumber*) key {
-    bool removed = false;
-    if([firestoreListeners objectForKey:key] != nil){
-        id<FIRListenerRegistration> firestoreListener = [firestoreListeners objectForKey:key];
-        [firestoreListener remove];
-        [firestoreListeners removeObjectForKey:key];
-        removed = true;
+    @synchronized (firestoreListeners) {
+        bool removed = false;
+        if([firestoreListeners objectForKey:key] != nil){
+            id<FIRListenerRegistration> firestoreListener = [firestoreListeners objectForKey:key];
+            [firestoreListener remove];
+            [firestoreListeners removeObjectForKey:key];
+            removed = true;
+        }
+        return removed;
     }
-    return removed;
 }
 
-/********************************/
+/*
+ * Functions
+ */
+- (void)functionsHttpsCallable:(CDVInvokedUrlCommand*)command {
+    [self.commandDelegate runInBackground:^{
+        @try {
+            __weak __auto_type weakSelf = self;
+            NSString* name = [command.arguments objectAtIndex:0];
+            NSDictionary* arguments = [command.arguments objectAtIndex:1];
+            [[[FIRFunctions functions] HTTPSCallableWithName:name] callWithObject:arguments
+                                                                  completion:^(FIRHTTPSCallableResult* _Nullable result, NSError* _Nullable error) {
+                if (error != nil) {
+                    [weakSelf sendPluginErrorWithError:error command:command];
+                } else {
+                    [weakSelf sendPluginDictionaryResult:result.data command:command callbackId:command.callbackId];
+                }
+            }];
+        }@catch (NSException *exception) {
+            [self handlePluginExceptionWithContext:exception :command];
+        }
+    }];
+}
+
+/*
+ * Installations
+ */
+- (void) getInstallationId:(CDVInvokedUrlCommand*)command {
+    [self.commandDelegate runInBackground:^{
+        @try {
+            [[FIRInstallations installations] installationIDWithCompletion:^(NSString *identifier, NSError *error) {
+                [self handleStringResultWithPotentialError:error command:command result:identifier];
+            }];
+        }@catch (NSException *exception) {
+            [self handlePluginExceptionWithContext:exception :command];
+        }
+    }];
+}
+
+- (void) getInstallationToken:(CDVInvokedUrlCommand*)command {
+    [self.commandDelegate runInBackground:^{
+        @try {            
+            [[FIRInstallations installations] authTokenForcingRefresh:true
+                                                           completion:^(FIRInstallationsAuthTokenResult *result, NSError *error) {
+              if (error != nil) {
+                  [self sendPluginErrorWithError:error command:command];
+              }else{
+                  [self sendPluginStringResult:[result authToken] command:command callbackId:command.callbackId];
+              }
+            }];
+        }@catch (NSException *exception) {
+            [self handlePluginExceptionWithContext:exception :command];
+        }
+    }];
+}
+
+- (void) deleteInstallationId:(CDVInvokedUrlCommand*)command {
+    [self.commandDelegate runInBackground:^{
+        @try {
+            [[FIRInstallations installations] deleteWithCompletion:^(NSError *error) {
+                [self handleEmptyResultWithPotentialError:error command:command];
+            }];
+        }@catch (NSException *exception) {
+            [self handlePluginExceptionWithContext:exception :command];
+        }
+    }];
+}
+
+- (void) sendNewInstallationId {
+    [self.commandDelegate runInBackground:^{
+        @try {
+            [[FIRInstallations installations] installationIDWithCompletion:^(NSString *identifier, NSError *error) {
+                if(error != nil){
+                    [self handlePluginErrorWithoutContext:error];
+                }else if(currentInstallationId != identifier){
+                    [FirebasePlugin.firebasePlugin executeGlobalJavascript:[NSString stringWithFormat:@"FirebasePlugin._onInstallationIdChangeCallback('%@')", identifier]];
+                    currentInstallationId = identifier;
+                }
+            }];
+        }@catch (NSException *exception) {
+            [self handlePluginExceptionWithoutContext:exception];
+        }
+    }];
+}
+
+/*************************************************/
 #pragma mark - utility functions
-/********************************/
+/*************************************************/
 - (void) sendPluginSuccess:(CDVInvokedUrlCommand*)command{
     [self.commandDelegate sendPluginResult:[CDVPluginResult resultWithStatus:CDVCommandStatus_OK] callbackId:command.callbackId];
 }
@@ -1930,6 +2040,10 @@ static NSMutableDictionary* firestoreListeners;
 - (void) sendPluginSuccessAndKeepCallback:(NSString*)callbackId{
     CDVPluginResult *pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK];
     [pluginResult setKeepCallbackAsBool:YES];
+}
+
+- (void) sendPluginNoResult:(CDVInvokedUrlCommand*)command callbackId:(NSString*)callbackId {
+    CDVPluginResult *pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_NO_RESULT];
     [self.commandDelegate sendPluginResult:pluginResult callbackId:callbackId];
 }
 
@@ -1939,9 +2053,30 @@ static NSMutableDictionary* firestoreListeners;
     [self.commandDelegate sendPluginResult:pluginResult callbackId:callbackId];
 }
 
+- (void) sendPluginStringResult:(NSString*)result command:(CDVInvokedUrlCommand*)command callbackId:(NSString*)callbackId {
+    CDVPluginResult *pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsString:result];
+    [self.commandDelegate sendPluginResult:pluginResult callbackId:callbackId];
+}
+
 - (void) sendPluginStringResultAndKeepCallback:(NSString*)result command:(CDVInvokedUrlCommand*)command callbackId:(NSString*)callbackId {
     CDVPluginResult *pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsString:result];
     [pluginResult setKeepCallbackAsBool:YES];
+    [self.commandDelegate sendPluginResult:pluginResult callbackId:callbackId];
+}
+
+- (void) sendPluginBoolResult:(BOOL)result command:(CDVInvokedUrlCommand*)command callbackId:(NSString*)callbackId {
+    CDVPluginResult *pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsBool:result];
+    [self.commandDelegate sendPluginResult:pluginResult callbackId:callbackId];
+}
+
+- (void) sendPluginBoolResultAndKeepCallback:(BOOL)result command:(CDVInvokedUrlCommand*)command callbackId:(NSString*)callbackId {
+    CDVPluginResult *pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsBool:result];
+    [pluginResult setKeepCallbackAsBool:YES];
+    [self.commandDelegate sendPluginResult:pluginResult callbackId:callbackId];
+}
+
+- (void) sendPluginDictionaryResult:(NSDictionary*)result command:(CDVInvokedUrlCommand*)command callbackId:(NSString*)callbackId {
+    CDVPluginResult *pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsDictionary:result];
     [self.commandDelegate sendPluginResult:pluginResult callbackId:callbackId];
 }
 
@@ -1978,7 +2113,7 @@ static NSMutableDictionary* firestoreListeners;
      if (error) {
          [self sendPluginErrorWithError:error command:command];
      }else{
-         [self.commandDelegate sendPluginResult:[CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsString:result] callbackId:command.callbackId];
+         [self sendPluginStringResult:result command:command callbackId:command.callbackId];
      }
 }
 
@@ -1986,7 +2121,7 @@ static NSMutableDictionary* firestoreListeners;
      if (error) {
          [self sendPluginErrorWithError:error command:command];
      }else{
-         [self.commandDelegate sendPluginResult:[CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsBool:result] callbackId:command.callbackId];
+         [self sendPluginBoolResult:result command:command callbackId:command.callbackId];
      }
 }
 
@@ -2000,6 +2135,11 @@ static NSMutableDictionary* firestoreListeners;
 - (void) handlePluginExceptionWithoutContext: (NSException*) exception
 {
     [self _logError:[NSString stringWithFormat:@"EXCEPTION: %@", exception.reason]];
+}
+
+- (void) handlePluginErrorWithoutContext: (NSError*) error
+{
+    [self _logError:[NSString stringWithFormat:@"ERROR: %@", error.description]];
 }
 
 - (void)executeGlobalJavascript: (NSString*)jsString